ferramentas/web/05_playwright/requirements.txt
pytest-playwright
allure-pytest